car/box: send telemetry data
- To be able to tune PID, car soft can send telemtry data now.
- Change a bit to template EventQueue class so it can adapt to the event struct.
- Rename motor A/B pins into DIR/SPEED
- Add pio GaspettoBox project, and add it to CI

+++ soft/emu-pc/arduino_framework/Arduino.cpp
@@ -1,15 +1,16 @@
 #include "Serial.h"
+#include "printf.h"
 
 #include <atomic>
 #include <chrono>
 #include <thread>
 
-std::atomic<unsigned long> millisCounter{ 0 };
+std::atomic<uint32_t> millisCounter{ 0 };
 std::atomic<bool> lowPowerMode;
 
-/*  Simulated millis function. */
 extern "C" {
-unsigned long millis(void)
+/*  Simulated millis function. */
+uint32_t millis(void)
 {
     return millisCounter.load();
 }
